// src/lib/api/generated/models/schemaDiagnosticOmission.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Hubuum REST API
 * OpenAPI documentation for the Hubuum REST service.
 * OpenAPI spec version: 0.0.15
 */

export type SchemaDiagnosticOmission = typeof SchemaDiagnosticOmission[keyof typeof SchemaDiagnosticOmission];


export const SchemaDiagnosticOmission = {
  actual_value_redacted: 'actual_value_redacted',
  instance_path_redacted_or_too_long: 'instance_path_redacted_or_too_long',
  schema_constraint_unavailable_or_too_large: 'schema_constraint_unavailable_or_too_large',
} as const;